requirements.txt
flask[async]>=3.0.0
# playwright>=1.53.0
requests==2.32.3
# supabase==2.10.0
dotenv==0.9.9
flask-cors==5.0.0
flask_sqlalchemy==3.1.1
gunicorn==23.0.0
flask_migrate==4.1.0
psycopg2-binary==2.9.10
PyJWT == 2.10.1
pytz == 2025.2
werkzeug==3.1.5
python-dateutil==2.9.0